requirements.txt
asgiref==3.3.1
certifi==2020.11.8
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
diff-match-patch==20200713
Django==2.0.2
django-cors-headers==3.5.0
django-crispy-forms==1.10.0
django-filter==2.4.0
django-formtools==2.2
django-guardian==2.3.0
django-import-export==2.4.0
django-reversion==3.0.8
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
et-xmlfile==1.0.1
future==0.18.2
httplib2==0.9.2
idna==2.10
importlib-metadata==3.1.1
itypes==1.2.0
jdcal==1.4.1
Jinja2==2.11.2
Markdown==3.3.3
MarkupPy==1.14
MarkupSafe==1.1.1
mysqlclient==2.0.1
odfpy==1.4.1
openpyxl==3.0.5
Pillow==8.0.1
PyJWT==1.7.1
PyMySQL==0.10.1
pytz==2020.4
PyYAML==5.3.1
requests==2.25.0
six==1.15.0
sqlparse==0.4.1
tablib==2.0.0
uritemplate==3.0.1
urllib3==1.26.2
xadmin @ https://codeload.github.com/sshwsfc/xadmin/zip/django2
xlrd==1.2.0
xlwt==1.3.0
zipp==3.4.0
